feat: widen stock quantities to Numeric + unit stamp (uom step 1)
Step 1 of the units-of-measure track (Refs #893). A widening and nothing
else: the seven quantity columns become Numeric(18,6) in Postgres, but
every schema, annotation and frontend gate above the database still says
`int`, so the API accepts and returns whole numbers exactly as before.

Migration 0074:
  - widens stock_entries.quantity_delta, project_entries.quantity,
    order_entries.quantity_{ordered,received}, lots.purchase_quantity,
    parts.low_stock_report_quantity and parts.attrition_min_quantity;
  - adds parts.unit_of_measure and an immutable per-row stock_entries.unit
    stamp, both String(16) NOT NULL DEFAULT 'pcs'. Stamping the unit on the
    row rather than resolving it through the part on read is what stops a
    part edit retroactively reinterpreting history.

integer -> numeric is lossless, so this stays reversible until the first
fractional value is written — and nothing here can write one. downgrade()
therefore refuses on any fractional or out-of-int4-range row rather than
letting Postgres' rounding numeric -> integer cast destroy it, mirroring
the guard 0032 put on its own upgrade.

The ALTERs rewrite their tables under ACCESS EXCLUSIVE; the migration sets
a 5s lock_timeout so it fails fast instead of queueing ahead of every
reader. stock_entries needs a maintenance window.

Application side is plumbing only: quantity_out() pins the untyped route
and MCP serialisers back to JSON integers (a scaled Decimal would render
as 5.0), and builds/service.py::_required coerces the now-Decimal
attrition floor so its float arithmetic stays bit-identical.

+++ backend/app/domain/projects/models.py
@@ -79,7 +79,11 @@ class ProjectEntry(WorkspaceOwned, Base):
     part_id = Column(UUID(as_uuid=True), ForeignKey("parts.id", ondelete="SET NULL"), nullable=True)
     meta_part_id = Column(UUID(as_uuid=True), ForeignKey("parts.id", ondelete="SET NULL"), nullable=True)
     name = Column(String(300), nullable=True)
-    quantity = Column(Integer, nullable=False, default=1)
+    # Numeric(18,6) since alembic 0074 (units-of-measure step 1) — this is
+    # the column migration 0032 narrowed to Integer for DB-005, widened back
+    # now that the ledger is Numeric too. BOM import still rejects fractional
+    # input; that inversion belongs to a later step of the track.
+    quantity = Column(Numeric(18, 6), nullable=False, default=1)
     # Per-BOM-line waste rate (Track B1, migration 0072). Inflates the
     # required + consumed quantity for a build; see
     # builds/service.py::_required. server_default keeps the NOT NULL add
